feat: Support incompatible schema changes
In this diff we build list of deltas of incompatible schema changes,
from the history of the collections schema changes.

On read and update we apply all the schema deltas, which has newer
version then the version of the row.

Incompatible schema changes supported:
  * field deletions
  * field type changes
  * reducing maxLength of strings and bytes

For the type changed we are doing best effort convertion.
If the type is not convertible we are returning empty value for it.
For the convertion rules please refer to `TestTypeConversionRules`.

Future work:
  * Currently the feature doesn't work if search is enabled.
    (turned out TypeSense doesn't allow field drop/add if data present
     in the collection)
  * Support max_length truncation for the strings
  * We can also support adding default value to the field. Populating
    empty fields in the existing rows with new default value.
    This can work for static defaults and updated_at(), created_at().

Feature can be configured by schema.allow_incompatible = true setting