fix: Secondary Index changes
chore: Enable secondary index writes for tests

fix: Ignore missing fields for arrrays fields

fix: Refactor the way the iterator was chosen for reading the docs.
There was a subtle bug because the variables were declared at the top.

fix: Add a separate table for secondary indexes so that it doesn't
interfere with scanning the primary key index and none of the index
writes get sent to the search indexer