feat: Manage secondary indexes via metadata
Manage the state of secondary indexes in the collections metadata.
Secondary indexes can be in 4 states:
    1. UNKNOWN = Have not fetched the index metadata yet so the state is unknown
	2. NOT_INDEXED = field is not indexed
	3. INDEX_WRITE_MODE = index is being built in the background and cannot be used for queries
	4. INDEX_ACTIVE = index can be used for queries

This allows Tigris to only use active and up to date indexes for
querying.

There is also code clean up to differentiate between the primary key
and the secondary index. The primary key index is kept in its own
metadata subspace.